==> ag_grant/04-OMB_counties.ipynb <==
NON-PROD

# default_exp OMB_counties

# The OMB urban concept has two categories: Metropolitan Areas and Micropolitan Areas, both created from counties
# according to criteria described here: 
# https://www.govinfo.gov/content/pkg/FR-2010-06-28/pdf/2010-15605.pdf.

# This notebook defines and maps counties that are
# metropolitan, micropolitan, and neither ("rural").

# export
import pandas as pd

# export
words = {'0':'rural','1':'micropolitan','2':'metropolitan'}
def make_word(code):
    """Apply a text description to the CBSA Level code."""
    return words[code]

def is_it_rural(code):
    """Create a 1/0 rural/urban flag."""
    if code == '0':
        return '1'
    else:
        return '0'

# export
# open a log file
logfile = open('004-OMB.log','a')

# export
for yr in range(1997,2018):
    print(f'\n{yr}:',file=logfile)
    data_file = f'/InfoGroup/data/rurality/InfoGroup_final_{yr}.csv'
    df = pd.read_csv(data_file,dtype=object)
    df['CBSA Level desc'] = df['CBSA Level'].apply(make_word)
    df['rural_OMB'] = df['CBSA Level'].apply(is_it_rural)
    print(f'rural/urban by OMB standard:')
    print(df['rural_OMB'].value_counts(),file=logfile)
    print(df['rural_OMB'].value_counts(normalize=True) * 100,file=logfile)
    # Write a new file
    outfile = f'/InfoGroup/data/rurality/InfoGroup_final_OMB_{yr}.csv'
    df.to_csv(outfile,index=None)

# export
logfile.close()
