==> ag_grant/05-points-in-polygons.ipynb <==
# default_exp Census_Bureau_urban_areas

# From https://www2.census.gov/geo/pdfs/reference/GARM/Ch12GARM.pdf:

# "Urbanized Areas (UAs)
# A UA is a continuously built-up area with a population of 50,000 or more.
# It comprises one or more places—central place(s)—and the adjacent densely settled surrounding 
# area—urban fringe—consisting of other places and nonplace territory."

# "Urban Places Outside of UAs
# Outside of UAs, an urban place is any incorporated place or census designated place (CDP) with 
# at least 2,500 inhabitants. A CDP is a densely settled population center that has a name and 
# community identity, and is not part of any incorporated place."

# "Rural Places and Territory
# Territory, population, and housing units that the Census Bureau does not classify as urban 
# are classified as rural."

# Census Urban Areas and Urban Clusters are not based on political boundaries at all.

# export
import pandas as pd
import numpy as np

# export
# Eventually add the Urban Area (UA) to each row by matching InfoGroup 'CBSA Code' 
# (metro/micropolitan area) to the Census Relationship file that cross-references CBSA and UA.
# The file contains a record for all metro/micro areas (CBSAs) and all urban areas/clusters 
# (UAs) in 2010.
# Prepare the data in the relationship file.
relationship = '/home/tflory/notebooks/InfoGroup/rurality/Relationship_Files/Urban_Area_to_Metro_Micro_Area_utf-8.csv'
rel = pd.read_csv(relationship)
cbsa_list = rel['CBSA'].drop_duplicates().tolist()
ua_list = rel['UA'].drop_duplicates().tolist()
cbsa_list.remove(99999)
ua_list.remove(99999)

# On the average there are is least a part of more than 3 UA's for each CBSA. Thus the
# only way to get a one-to-one mapping of CBSA to UA will be to do the shapely 'within' 
# comparison; determine whether the point coordinates of an IG enterprise fall within 
# the polygon coordinates of a particular UA.

# But to create a simple 'rural_Census' flag we need only to discover whether the CBSA Code 
# of the IG record matches any CBSA code in the relationship file that is not in an urban area.

# export
for yr in range(1997,2018):
    infile = f'/InfoGroup/data/rurality/InfoGroup_final_OMB_{yr}.csv'
    df = pd.read_csv(infile)
    df['CBSA Code'] = df['CBSA Code'].fillna(99999)
    df['CBSA Code'] = df['CBSA Code'].astype(int)


# To use a df copy instead of reading it in again. Uncomment the useful line.
#df_copy = df.copy()
#df = df_copy.copy()

print(len(df[df['CBSA Code']==0]))
print(len(df[df['CBSA Code']==99999]))

len(df['CBSA Code'].drop_duplicates())


print(len(cbsa_list))
print(len(ua_list))

# The file contains a record for all metro/micro areas (CBSAs) and all urban areas/clusters (UAs) in 2010.
# Exploration:
print(len(rel))
print(len(rel[['UA','CBSA']].drop_duplicates()), '\n')
# There are 5167 records in the file, all with paired values of UA and CBSA that are unique.

print(len(rel[['UA','CBSA']][rel['CBSA']==99999]))
print(len(rel[['UA','CBSA']][rel['CBSA']!=99999]),'\n')
# There are 1038 records with a valid UA code but NO metro/micro area (CBSA).
# There are 4129 UAs with a valid CBSA.

print(len(rel[['UA','CBSA']][rel['UA']==99999]))
print(len(rel[['UA','CBSA']][rel['UA']!=99999]),'\n')
# There are 955 records with a valid CBSA code but NO urban area/cluster (UA).
# There are 4212 CBSAs with a valid UA.

print(len(rel[(rel['CBSA']!=99999) & (rel['UA']!=99999)]))
# There are 3174 records with valid unique combinations of the CBSA and UA values.

print(len(rel[(rel['CBSA']==99999) & (rel['UA']==99999)]))
# By definition there are no cases in which both the CBSA and UA are missing.

print(len(rel['CBSA'].drop_duplicates()))
print(len(rel['UA'].drop_duplicates()))

# ----------------------------

# Is a UA in a CBSA 1) partly, 2) entirely, or 3) not at all?
# Make a dict with the UA as the key and the value a list of all CBSA codes.

# The uas dict has a list of CBSAs for each UA. The ua_keys dict tags each UA as rural, partly rural, or
# urban. But in InfoGroup we have only CBSA Code. We can determine if the IG record's CBSA Code matches to a UA that 
# is entirely urban and that is completely within the CBSA; i.e., it has only the IG record's CBS code in its
# CBSA list. And if it matches to a UA that is entirely rural; i.e., the UA has only a single 99999 in its CBSA list.

# The cbsas dict is a reverse version of uas: a key for every CBSA and, as the value, the list of UAs associated
# with the key. The cbsas_keys dict substitutes the urban/rural tag of each UA for its code value. Using those 
# lists of text values we can determine whethera record is rural, partly rural, or urban.

# UA key

# Create dict from list of lists.
uas = {}
for l in rel[['UA','CBSA']].values.tolist():
    try:
        uas[l[0]].append(l[1])
    except:
        uas[l[0]] = [l[1]]

# Validate that 99999 can occur only once in any value list in the uas dict.
n = 0
m = 0
for k,v in uas.items():
    if v.count(99999) > 1:
        n += 1
    elif v.count(99999) == 1:
        m += 1
        
print(n,m)

# Categorize all the UAs as urban, rural, partly rural, or unknown.
# Partly rural == UA with some valid CBSAs and at least 1 99999;
# Rural (entirely rural) == UA with no valid CBSAs, just some 99999s;
# Urban (not at all rural) == UA with some valid CBSAs but no 99999s.
# Unknown == UA 99999 (missing)

def all_nines(li):
    length = len(li)
    n = 0
    for item in li:
        if item == 99999:
            n += 1
    if n == length:
        return True
    return False

ua_keys = dict.fromkeys(uas.keys())
for k,v in uas.items():
    if len(v) > 1:
        if all_nines(v):
            ua_keys[k] = 'rural-multi'
        elif 99999 in v:
            ua_keys[k] = 'partly rural'
        elif k == 99999:
            ua_keys[k] = 'unknown'
        else:
            ua_keys[k]= 'urban-multi'
    else:
        if v[0] == 99999:
            ua_keys[k] = 'rural-single'
        else:
            ua_keys[k] = 'urban-single'

# Validate that all keys have values [and remove the key for UA 99999].
for i,v in ua_keys.items():
    if v == None: 
        print(i)
print(ua_keys[99999])

# No, don't delete the key for 99999
#del ua_keys[99999]
#try:
#    print(ua_keys[99999])
#except KeyError:
#    print('No such key')

len_att = len(ua_keys)
entire = 0
urban = 0
part = 0
unk = 0
for k,v in ua_keys.items():
    if v in ['rural-single','rural-multi']:
        entire += 1
    elif v == 'partly rural':
        part += 1
    elif v in ['urban-single','urban-multi']:
        urban += 1
    elif v == 'unknown':
        unk += 1

print('Out of',str(len_att),'UAs:','\n',str(entire),'are entirely rural,','\n',str(part),'are partly rural,','\n',
      str(urban),'are urban, and','\n',str(unk),'is unknown.')

print(str((2563/3602)*100))
print(str((908/3602)*100))
print(str((130/3602)*100))
print(str((1/3602)*100))

# CBSA key
# a CBSA is not covered entirely by UAs, so is it valid to assess the rurality of a record if it only has a
# CBSA Code and no way to categorize the parts of the CBSA that are not covered by any UA?

# Create dict from list of lists.
cbsas = {}
for l in rel[['UA','CBSA']].values.tolist():
    try:
        cbsas[l[1]].append(l[0])
    except:
        cbsas[l[1]] = [l[0]]

# function to remove all occurrences of a particular value from a list.
def remove_all(li,val):
    i = 0
    while i < len(li):
        try:
            li.remove(val)
        except:
            pass
        i += 1
    return li

del cbsas[99999]
for k,v in cbsas.items():
    cbsas[k] = remove_all(v,99999)

len(cbsas)

# Find the list of UAs that are associated with a single CBSA code.
def find_uas(cbsa): 
    uas_list = []
    for k in uas.keys():
        #if k == 99999: 
            #continue
        if cbsa in uas[k]:
            uas_list.append(k)
    return uas_list

# Get a list of the urban/rural descriptions for each UA in a list of UAs.
def find_descs(uas_list):
    descs_list = []
    for ua in uas_list:
        descs_list.append(ua_keys[ua])
    return descs_list

# Object:
#       cbsa_keys: a dict with CBSA codes as keys and empty lists to be filled in with urban/rural text strings as values.

cbsa_keys = dict.fromkeys(cbsas.keys())

# Data: 
#       uas: a dict with UA codes as keys and lists of constituent CBSAs as values.
#       uas_keys: a dict with UA codes as keys and one of 'rural','partly rural', and 'urban' as a value.
#       cbsas: a dict with CBSA values as keys and lists of constituent UAs as values.

for cbsa in cbsa_keys.keys():
    li = find_uas(cbsa)
    cbsa_keys[cbsa] = find_descs(li)

# Validation
print(cbsas[10020])
print(cbsa_keys[10020])
print(ua_keys[37],ua_keys[43993],ua_keys[46045],ua_keys[99999])

# Validation
print(cbsas[48140])
print(cbsa_keys[48140])
print(ua_keys[91],ua_keys[55144],ua_keys[93025],ua_keys[99999])

# Now add an 'invalid_rural_Census' variable with appropriate values(?). Invalid because the logic is
# faulty, as it turns out. For example...
# If one description is 'partly rural', the whole CBSA is partly rural
# If there is a mixture of urban and rural (and partly rural), the whole CBSA is partly rural
# Otherwise, rural and urban are both all rural or all urban
def assign_text(code):
    try:
        li = cbsa_keys[code]
    except KeyError:
        return 'not available'
    
    if 'partly rural' in li:
        return 'partly rural'
    elif 'urban' in li and 'rural' in li:
        return 'partly rural'
    elif 'rural' in li:
        return 'rural'
    elif 'urban' in li:
        return 'urban'
    else:
        return 'unknown'

df['invalid_rural_Census'] = df['CBSA Code'].apply(assign_text)

df['invalid_rural_Census'].value_counts()

df['invalid_rural_Census'].value_counts(normalize=True) * 100

# No cases in which all UAs for a CBSA are rural?
n = 0
for i,v in cbsa_keys.items():
    if v.count('rural') == len(v):
        n += 1
n

# Are all the 'not availables' KeyErrors? Yes.
unknowns = set([i for i in df['CBSA Code'][df['invalid_rural_Census']=='not available']])
print(len(unknowns))

n = 0
master = set(cbsas.keys())
for unk in unknowns:
    if unk in master:
        n += 1
print(n)

# Thus 53 CBSA Codes in the InfoGroup file are not listed in the relationship file. These are the CBSA Code
# values for 1,691,022 records.

# Simple 'reverse-rurality' indicator: for each CBSA in the InfoGroup file, does it match to any CBSA in the 
# relationship file that is matched there to a valid UA code?

pairs = rel[['UA','CBSA']].values.tolist()
print(len(pairs))
urban_cbsas = []
for pair in pairs:
    if pair[0] == 99999 or pair[1] == 99999:
        continue
    else:
        urban_cbsas.append(pair[1])
print(len(urban_cbsas))
urban_cbsas_set = set(urban_cbsas)
print(len(urban_cbsas_set))

def isit_rural(code):
    if code in urban_cbsas_set:
        return 0
    else:
        return 1

df['rural_Census_general'] = df['CBSA Code'].apply(isit_rural)

print(df['rural_Census_general'].value_counts())
print(df['rural_Census_general'].value_counts(normalize=True) * 100)

df.drop('invalid_rural_Census',axis=1,inplace=True)

# ---------------------------------

# For later analysis we will want the six-digit NAICS codes as well as the 8-digit code that comes with
# InfoGroup and the 2-digit code and description that we created in the #1 notebook.
# We'll also include some descriptions for some codes that the project statement mentions specifically as
# definitely agricultural.

df['NAICS6'] = df['Primary NAICS Code'].apply(lambda x: str(x)[:6])

ag_naics = {'112112':'Cattle Feedlots',
            '115111':'Cotton Ginning',
            '115112':'Soil Preparation, Planting and Cultivating',
            '115113':'Crop Harvesting, Primarily by Machine',
            '115114':'Postharvest Crop Activities except Cotton Ginning',
            '115115':'Farm Labor Contractors and Crew Leaders',
            '115116':'Farm Management Services',
            '423820':'Farm and Garden Machinery and Equipment Merchant Wholesalers',
            '424430':'Dairy Product Merchant Wholesalers',
            '424440':'Poultry and Poultry Product Merchant Wholesalers',
            '424480':'Fresh Fruit and Vegetable Merchant Wholesalers',
            '424510':'Grain and Field Bean Merchant Wholesalers',
            '424520':'Livestock Merchant Wholesalers',
            '424590':'Other Farm Product Raw Material Merchant Wholesalers',
            '424910':'Farm Supplies Merchant Wholesalers',
            '493130':'Farm Product Warehousing and Storage'}

def naics6_desc(code):
    if code in ag_naics.keys():
        return ag_naics[code]
    return np.nan

df['NAICS6 desc'] = df['NAICS6'].apply(naics6_desc)

df['NAICS6 desc'].value_counts()

len(df[df['NAICS6 desc'].isnull()])

df.columns

outfile = 'data/df_%d_OMB_Census_TEMP.csv' % year
df.to_csv(outfile,index=None)

# How to do the point-in-polygon comparison...

%%time
year = 2017
infile = 'data/df_%d_OMB_Census_TEMP.csv' % year
df = pd.read_csv(infile,dtype=object)

df[['Latitude','Longitude']] = df[['Latitude','Longitude']].apply(pd.to_numeric)

dfx = df[['ABI','Latitude','Longitude']].sample(n=10000)

def pointer(lon,lat):
    return Point(lon,lat)

%%time
dfx['geometry'] = dfx.apply(lambda x: pointer(x['Longitude'], x['Latitude']), axis=1)

dfx.drop(columns=['Latitude','Longitude'],inplace=True)

The 2010 Urban Area to Metropolitan and Micropolitan Statistical Areas Relationship File provides relationships between the 2010 urban areas and the Metro and Micro Areas. Urban areas include both urbanized areas (UA) and urban clusters (UC), and areas outside urban areas are considered rural. 

This list can be utilized to see if a metro or micro area is urban, rural, or contains both urban and rural area.

Each record in the relationship file represents one polygon that is formed when the 2010 urban areas overlay the metro and micro areas. The polygon may represent the intersection of a 2010 urban area and a metro or micro area, or the portion of the metro or micro area that does not contain a 2010 urban area. In addition, since metro and micro areas are not a wall-to-wall coverage across the U.S., the polygon may represent the area of a 2010 urban area that does not intersect a metro and micro area.

# The polygons of the UAs.
shape = 'map_files/tl_2017_us_uac10.shp'
ua = gpd.read_file(shape)

#ua.columns

ua.set_index("UACE10", inplace = True)

ua_dict = dict.fromkeys(ua.index)
for k in ua_dict.keys():
    ua_dict[k] = ua.loc[k]['geometry']

def set_ua(pnt):
    for k,p in ua_dict.items():
        if pnt.within(p):
            return k
    return '99999'

%%time
dfx['UA'] = dfx.apply(lambda row: set_ua(row['geometry']),axis=1)

# vectorized execution of a function that takes 4 params. faster than .apply()
#df['distance'] = haversine(40.671, -73.985, df['latitude'], df['longitude'])
# even better: use numpy arrays produced under the hood by .values
#df['distance'] = haversine(40.671, -73.985, df['latitude'].values, df['longitude'].values)

# The 15 programs in the points-in-polygons folder ran simultaneously as a brute force method of parallel 
# processing to locate the InfoGroup points in UA polygons. They incorporate the logic, maybe not the exact code,
# of the non-executable cells just above. They produced 15 csv files with the UA code for
# the UA record along with the shapely geometry and the ABI of the record.  Now combine those 15 files into 
# one and merge the result into the df dataframe.

%%time
pp_df = pd.DataFrame()

for seq in range(1,16):
    infile = 'points-in-polygons/data/dfx_%d.csv' % seq
    df_tmp = pd.read_csv(infile,dtype=object)
    pp_df = pd.concat([pp_df,df_tmp])

print(len(pp_df))

pp_df['ABI'] = pp_df['ABI'].astype(int)

%%time
# Merge with df
df_final = df.merge(pp_df,on='ABI',how='inner',indicator=True)
df_final.columns

df_final['_merge'].value_counts()

# The polygon 'geometry' is no longer necessary. Ditto for '_merge'.
df_final.drop(columns=['geometry','_merge'],inplace=True)

# Create the definitive 'rural_Census' variable.

# Informational
rural_pct = len(df_final[df_final['UA'] == '99999']) / len(df_final) * 100.0
print('Rural percentage by the most specific Census-based standard',str(rural_pct))

df_final['rural_Census'] = df_final['UA'].apply(lambda x: 1 if x == '99999' else 0)

df_final['rural_Census'].value_counts(normalize=True) * 100.0

# In the InfoGroup file we now have the Urban Area identifier. That can be merged with locational variables from 
# the Census shapefile for mapping. The output file here is a regular .csv file, not a shapefile, but it does have
# identifiers and coordinates for each UA and can be read into a GeoDataFrame.

df_final.columns

%%time
outfile = 'data/df_%d_OMB_Census.csv' % year
df_final.to_csv(outfile,index=None)
