==> ag_grant/12-Employment Summary.ipynb <==
import pandas as pd

infile = 'data/df_2017_OMB_Census_HRSA_FAR.csv'
df = pd.read_csv(infile,dtype=object)

df.dtypes

df['Employee Size (5) - Location'] = df['Employee Size (5) - Location'].astype(float)
total_emp = df['Employee Size (5) - Location'].sum()
print("Total employment:",str(total_emp))

total_OMB_rural = df[df['rural_OMB']=='1']['Employee Size (5) - Location'].sum()
pct = (total_OMB_rural / total_emp) * 100.0
print('OMB rural percentage employment:',str(pct))

total_Census_rural_general = df[df['rural_Census_general']=='1']['Employee Size (5) - Location'].sum()
pct = (total_Census_rural_general / total_emp) * 100.0
print('Workaround Census rural percentage employment:',str(pct))

total_Census_rural = df[df['rural_Census']=='1']['Employee Size (5) - Location'].sum()
pct = (total_Census_rural / total_emp) * 100.0
print('Precise Census rural percentage employment:',str(pct))

total_HRSA_rural = df[df['rural_HRSA']=='1']['Employee Size (5) - Location'].sum()
pct = (total_HRSA_rural / total_emp) * 100.0
print('HRSA rural percentage employment:',str(pct))

# For FAR statistics, include only records from the lower 48 without missing 'FAR Level' values.
df['FAR Level'].fillna(-1,inplace=True)
df[['Continental','State FIPS','FAR Level']] = df[['Continental','State FIPS','FAR Level']].astype(float)
df[['Continental','State FIPS','FAR Level']] = df[['Continental','State FIPS','FAR Level']].astype(int)
df_select = df[(df['Continental']==1) & (df['State FIPS'] != 11) & (df['FAR Level'] >= 0)].copy()
total_FAR_emp = df_select['Employee Size (5) - Location'].sum()

total_FAR_rural = df_select[df_select['FAR Level'].astype(int) > 0]['Employee Size (5) - Location'].sum()
pct = (total_FAR_rural / total_FAR_emp) * 100.0
print('FAR (any level) rural percentage employment:',str(pct))
